Fix creating directories and copying files
Since both `:create_dirs` and `:copy_files` default to `[]`, that will already be present in `cfg` and take precedence over the items concatenated to the list. Prepending them to the list fixes the issue, and doesn't ignore user defined values either, since those are taken into consideration earlier in the function.